7/3

+++ modeling/speech_encoder.py
@@ -35,7 +35,7 @@ class SpeechEncoder(nn.Module):
             self.processor.feature_extractor.sampling_rate / 50
         )
         self.padding_length = 320
-        self.model = AutoModel.from_pretrained(model_id,torch_dtype=torch.float16)
+        self.model = AutoModel.from_pretrained(model_id)
         self.model_output_dim = self.model.config.hidden_size
         self.downsample_K = downsample_K
         self.project_dim = project_dim
